requirements.txt
Django==5.0.4
djangorestframework
djangorestframework-simplejwt
django-cors-headers
requests
transformers
